Added toast, confirm password
Added confirm password field on signup form.
Implemented toast message
(from https://github.com/apvarun/toastify-js)
I have loaded it through cdn and only on signup page but if you are happy with it we can add it through npm.

Did a small refactor on signup js so that we do not have to add a separate error handler for each status code